Fixed proxy authentication for HTTP
Fixes https://github.com/NuGet/Home/issues/2839 in 3.4.5.

This change is to handle proxy authenticaton response correctly for HTTP.
Originally only HTTPs response was handled with try/catch of an exception
of particular type.
Shared proxy credentials acquisition logic moved to a private method and
it's called now in both cases.

This bug was fixed in dev and 3.5.x branches earlier.